# Fix the tests on Windows