--- a/test/issues/940.test.js
+++ b/test/issues/940.test.js
@@ -160,7 +160,7 @@ describe("#940", () => {
       await promise2;
 
       background.browser.tabs.create.should.have.been.calledTwice;
-    }).timeout(2002);
+    }).timeout(2500);
 
     it("should not influence the canceled url in other tabs", async () => {
       const newTab = {