The tests really expect the files to be checked out with LF-only line endings. Let's declare that!

And also, give one test a bit more time, as it takes a little longer on Windows (probably due to file I/O).