@groovecoder 2019-04-15
Thanks for the PR @dscho . I don't have a Windows machine (nor VM) to test this ATM. @stoically or @jonathanKingston - do you?

If not I'll dig one up to double-check this.