@stoically 2019-04-15
The timeout fix in this PR is actually not necessary anymore after the refactoring in #1190 which gets rid of actual timeouts and replaces them with sinon fake timers. The eol=ef for gitattributes looks good to me.