@dscho 2019-04-17
> I don't have a Windows machine (nor VM) to test this ATM.

@groovecoder well, well, here you are: https://github.com/mozilla/multi-account-containers/pull/1374