@dscho 2019-04-18
> Could you try to execute `npm run test-watch` and see whether the tests without nyc are running?

Yep:

```
PS C:\git-sdk-64\usr\src\git\3rdparty\multi-account-containers> npm run test-watch

> testpilot-containers@6.1.0 test-watch C:\git-sdk-64\usr\src\git\3rdparty\multi-account-containers
> mocha ./test/setup.js test/**/*.test.js --watch




  Assignment Feature
    click the 'Always open in' checkbox in the popup
      open new Tab with the assigned URL in the default container
        √ should open the confirm page
        √ should remove the new Tab that got opened in the default container
      click the 'Always open in' checkbox in the popup again
        open new Tab with the no longer assigned URL in the default container
          √ should not open the confirm page

  External Webextensions
    with contextualIdentities permissions
      √ should be able to get assignments
    without contextualIdentities permissions
Error: Missing contextualIdentities permission
    at browser.runtime.onMessageExternal.addListener (file:///C:/git-sdk-64/usr/src/git/3rdparty/multi-account-containers/src/js/background/messageHandler.js:80:17)
      √ should throw an error

  #940
    when other onBeforeRequestHandlers are faster and redirect with the same requestId
      √ should not open two confirm pages (80ms)
    when redirects change requestId midflight
      √ should not open two confirm pages
      √ should uncancel after webRequest.onCompleted
      √ should uncancel after webRequest.onErrorOccurred
      √ should uncancel after 2 seconds (2004ms)
      √ should not influence the canceled url in other tabs


  11 passing (3s)
```